fixed the interface for scene
