+++ src/main.rs
@@ -7,7 +7,7 @@ mod common;
 mod scene;
 
 use canvas::Canvas;
-use common::{DEFAULT_SCREEN_HEIGHT, DEFAULT_SCREEN_WIDTH};
+use common::{Spectrum, DEFAULT_SCREEN_HEIGHT, DEFAULT_SCREEN_WIDTH, EPS};
 use scene::{Point, Ray, Scene, Vector};
 
 struct Config {
@@ -45,7 +45,7 @@ impl Raytracer {
         }
     }
 
-    fn render(&mut self) {
+    pub fn render(&mut self) {
         // start rendering
         // ray casting algorithm
         let x_width = 2.0 * f64::tan(self.config.fov / 2.0);
@@ -61,10 +61,51 @@ impl Raytracer {
                 let x_component = x_start + x_step * (i as f64);
                 let y_component = y_start - y_step * (j as f64);
                 let vector = Vector::new(x_component, y_component, -1.0);
-                let ray = Ray::new(self.config.origin, vector, 0);
-                let color = self.scene.cast_ray(&ray);
+                let ray = Ray::new(self.config.origin, vector);
+                let color = self.cast_ray(&ray, 0);
                 self.canvas.draw_pixel(i, j, color);
             }
+        }
+    }
+
+    fn cast_ray(&self, ray: &Ray, bounces_left: u32) -> Spectrum {
+        match self.scene.intersect(ray) {
+            Some(ray_intersection) => {
+                let object = ray_intersection.object;
+                let min_dist = ray_intersection.distance;
+                let mut intersection_point: Point = ray.get_intersection_point(min_dist);
+                // bumping the point a little out of the object to prevent self-collision
+                let surface_normal: Vector = object.surface_normal(intersection_point);
+                intersection_point = intersection_point + surface_normal.scale(EPS);
+
+                let emittance = object.material().emittance;
+                match bounces_left {
+                    0 => {
+                        // zero bounce radiance
+                        emittance
+                    }
+                    1 => {
+                        // direct lighting
+                        unimplemented!()
+                    }
+                    _ => {
+                        // global illumination
+                        unimplemented!()
+                    }
+                }
+                // for point_light in self.lights.iter() {
+                //     let light_direction = (point_light.position - intersection_point).normalized();
+                //     let light_ray = Ray::new(intersection_point, light_direction);
+
+                //     if self.object_intersection(&light_ray).is_none() {
+                // 	// lambertian code
+                // 	let intensity = f64::abs(light_direction.dot(surface_normal));
+                // 	let color_value = (intensity * 255.0) as u8;
+                // 	color += Spectrum::new(color_value, color_value, color_value);
+                //     }
+                // }
+            }
+            None => Spectrum::new(0, 0, 0),
         }
     }
