+++ src/scene/mod.rs
@@ -14,8 +14,8 @@ pub struct Scene {
 }
 
 pub struct RayIntersection<'a> {
-    distance: f64,
-    object: &'a dyn Object,
+    pub distance: f64,
+    pub object: &'a dyn Object,
 }
 
 impl Scene {
@@ -75,7 +75,7 @@ impl Scene {
         self.planes.len() + self.spheres.len()
     }
 
-    fn object_intersection(&self, ray: &Ray) -> Option<RayIntersection> {
+    pub fn intersect(&self, ray: &Ray) -> Option<RayIntersection> {
         let mut min_dist = f64::INFINITY;
         let mut min_object = None;
         for i in 0..self.num_objects() {
@@ -93,49 +93,6 @@ impl Scene {
                 distance: min_dist,
             }),
             None => None,
-        }
-    }
-
-    // for a ray, estimate global illumination
-    pub fn cast_ray(&self, ray: &Ray) -> Spectrum {
-        match self.object_intersection(ray) {
-            Some(ray_intersection) => {
-                let object = ray_intersection.object;
-                let min_dist = ray_intersection.distance;
-                let mut intersection_point: Point = ray.get_intersection_point(min_dist);
-                // bumping the point a little out of the object to prevent self-collision
-                let surface_normal: Vector = object.surface_normal(intersection_point);
-                intersection_point = intersection_point + surface_normal.scale(EPS);
-
-                let bounces_left = ray.bounces_left;
-                let emittance = object.material().emittance;
-                match bounces_left {
-                    0 => {
-                        // zero bounce radiance
-                        emittance
-                    }
-                    1 => {
-                        // direct lighting
-                        unimplemented!()
-                    }
-                    _ => {
-                        // global illumination
-                        unimplemented!()
-                    }
-                }
-                // for point_light in self.lights.iter() {
-                //     let light_direction = (point_light.position - intersection_point).normalized();
-                //     let light_ray = Ray::new(intersection_point, light_direction);
-
-                //     if self.object_intersection(&light_ray).is_none() {
-                // 	// lambertian code
-                // 	let intensity = f64::abs(light_direction.dot(surface_normal));
-                // 	let color_value = (intensity * 255.0) as u8;
-                // 	color += Spectrum::new(color_value, color_value, color_value);
-                //     }
-                // }
-            }
-            None => Spectrum::new(0, 0, 0),
         }
     }
 }
